Перенесёт загрузчик в PrivateRoute
