+++ src/components/private-route/private-route.jsx
@@ -1,11 +1,24 @@
-import React from 'react';
+import React, {useEffect} from 'react';
+import {connect} from 'react-redux';
 import PropTypes from 'prop-types';
 import {Route, Redirect} from 'react-router-dom';
-import {connect} from 'react-redux';
 import {AuthorizationStatus} from '../../const';
+import {fetchOffersList} from '../../store/api-actions';
+import LoadingScreen from '../loading-screen/loading-screen';
 
 
-const PrivateRoute = ({render, path, exact, authorizationStatus}) => {
+const PrivateRoute = ({render, path, exact, authorizationStatus, onLoadOffers, isOffersLoaded}) => {
+  useEffect(() => {
+    if (!isOffersLoaded) {
+      onLoadOffers();
+    }
+  }, [isOffersLoaded]);
+
+  if (!isOffersLoaded) {
+    return (
+      <LoadingScreen />
+    );
+  }
   return (
     <Route
       path={path}
@@ -26,12 +39,20 @@ PrivateRoute.propTypes = {
   exact: PropTypes.bool.isRequired,
   path: PropTypes.string.isRequired,
   render: PropTypes.func.isRequired,
+  isOffersLoaded: PropTypes.bool,
+  onLoadOffers: PropTypes.func,
 };
 
 const mapStateToProps = (state) => ({
   authorizationStatus: state.authorizationStatus,
+  isOffersLoaded: state.isOffersLoaded
 });
 
+const mapDispatchToProps = (dispatch) => ({
+  onLoadOffers() {
+    dispatch(fetchOffersList());
+  }
+});
 
 export {PrivateRoute};
-export default connect(mapStateToProps)(PrivateRoute);
+export default connect(mapStateToProps, mapDispatchToProps)(PrivateRoute);
